pyproject.toml
[tool.poetry]
name = "pgdatacleaner"
version = "1.1.0a16"
description = "Data cleaning/masking for PostgreSQL databases"
authors = ["Fredrik Håård <fredrik@metallapan.se>"]
license = "MIT License"

[tool.poetry.dependencies]
python = ">=3.7,<4"
psycopg2-binary = ">2.8.0,<2.10"
Faker = "^8.12.1"
python-slugify = "^5.0.2"

[tool.poetry.dev-dependencies]
tox = "^3.24.3"
tox-poetry = "^0.4.1"
pytest = "^6.2.5"
shed = "^0.3.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dataclean = 'pgdatacleaner.cleaner:main'
datadict = 'pgdatacleaner.datadict:main'
